requirements.txt
Django~=5.2
pandas
sentence-transformers
faiss-cpu
numpy
openai